# backend-api/docs/openapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chemistry Game Suite API",
    "description": "API cho bộ sưu tập trò chơi hóa học dựa trên OpenAPI 3.0, hỗ trợ quản lý phản ứng, ô chữ, nguyên tố, và phiên chơi. \nTìm hiểu thêm về Swagger tại [https://swagger.io](https://swagger.io).",
    "termsOfService": "https://yourchemistrygame.com/terms/",
    "contact": {
      "email": "giaphuc2004hg@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.12"
  },
...
  "servers": [
    {
      "url": "https://ct313hm01-project-gphuc04.onrender.com/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Lỗi dữ liệu đầu vào"
          },
          "error": {
            "type": "string",
            "example": "Invalid input"
          }
        },
        "required": ["code", "message"]
      },
...
    }
  },
  "paths": {
...
    "/games": {
      "get": {
        "tags": ["Games"],
        "summary": "Lấy danh sách tất cả các trò chơi",
        "description": "API để lấy danh sách tất cả các trò chơi, yêu cầu xác thực JWT.",
        "operationId": "getAllGames",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách trò chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách trò chơi thành công"
                    },
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "game_id": {
                            "type": "string",
                            "example": "game123"
                          },
                          "name": {
                            "type": "string",
                            "example": "Chemistry Quiz"
                          },
                          "difficulty_level": {
                            "type": "string",
                            "example": "Easy"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Games"],
        "summary": "Tạo mới một trò chơi",
        "description": "API để tạo một trò chơi mới, yêu cầu xác thực JWT. Hỗ trợ dữ liệu dạng multipart.",
        "operationId": "createGame",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "game_code": {
                    "type": "string",
                    "example": "master_chef",
                    "description": "Mã trò chơi duy nhất"
                  },
                  "name_vi": {
                    "type": "string",
                    "example": "Đầu Bếp Hóa Học",
                    "description": "Tên trò chơi bằng tiếng Việt"
                  },
                  "name_en": {
                    "type": "string",
                    "example": "Chemistry Chef",
                    "description": "Tên trò chơi bằng tiếng Anh"
                  },
                  "description_vi": {
                    "type": "string",
                    "example": "Trò chơi mô phỏng phòng thí nghiệm hóa học",
                    "description": "Mô tả trò chơi bằng tiếng Việt"
                  },
                  "description_en": {
                    "type": "string",
                    "example": "A chemistry lab simulation game",
                    "description": "Mô tả trò chơi bằng tiếng Anh"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "reaction_game",
                      "periodic_table",
                      "lab_simulation",
                      "puzzle_knowledge"
                    ],
                    "example": "lab_simulation",
                    "description": "Danh mục trò chơi"
                  },
                  "max_score": {
                    "type": "integer",
                    "example": 1000,
                    "description": "Điểm tối đa của trò chơi"
                  },
                  "time_limit": {
                    "type": "integer",
                    "example": 300,
                    "description": "Thời gian giới hạn của trò chơi (giây)"
                  },
                  "is_active": {
                    "type": "boolean",
                    "example": true,
                    "description": "Trạng thái hoạt động của trò chơi"
                  },
                  "game_settings": {
                    "type": "string",
                    "example": "{\"question_count\": 10, \"hints_allowed\": true}",
                    "description": "Cài đặt riêng của trò chơi dưới dạng chuỗi JSON"
                  }
                },
                "required": ["game_code", "name_vi", "name_en"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo trò chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "POST, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo trò chơi thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "game_id": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi dữ liệu đầu vào không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/games/{id}": {
      "get": {
        "tags": ["Games"],
        "summary": "Lấy thông tin trò chơi theo ID",
        "description": "API để lấy thông tin chi tiết của một trò chơi dựa trên gameId, yêu cầu xác thực JWT.",
        "operationId": "getGameById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID của trò chơi cần lấy thông tin",
            "schema": {
              "type": "string",
              "example": "game123"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin trò chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "GET, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thông tin trò chơi thành công"
                    },
                    "result": {
                      "type": "object",
                      "properties": {
                        "game_id": {
                          "type": "string",
                          "example": "game123"
                        },
                        "game_code": {
                          "type": "string",
                          "example": "master_chef"
                        },
                        "name_vi": {
                          "type": "string",
                          "example": "Bếp Trưởng Bậc Thầy"
                        },
                        "name_en": {
                          "type": "string",
                          "example": "Master Chef"
                        },
                        "category": {
                          "type": "string",
                          "example": "reaction_game"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID trò chơi không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trò chơi",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Games"],
        "summary": "Cập nhật trò chơi theo ID",
        "description": "API để cập nhật thông tin trò chơi dựa trên gameId, yêu cầu xác thực JWT. Hỗ trợ dữ liệu dạng multipart.",
        "operationId": "updateGame",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID của trò chơi cần cập nhật",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "game_code": {
                    "type": "string",
                    "maxLength": 20,
                    "nullable": true,
                    "description": "Mã định danh duy nhất của trò chơi"
                  },
                  "name_vi": {
                    "type": "string",
                    "maxLength": 100,
                    "nullable": true,
                    "description": "Tên trò chơi bằng tiếng Việt"
                  },
                  "name_en": {
                    "type": "string",
                    "maxLength": 100,
                    "nullable": true,
                    "description": "Tên trò chơi bằng tiếng Anh"
                  },
                  "description_vi": {
                    "type": "string",
                    "nullable": true,
                    "description": "Mô tả trò chơi bằng tiếng Việt"
                  },
                  "description_en": {
                    "type": "string",
                    "nullable": true,
                    "description": "Mô tả trò chơi bằng tiếng Anh"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "reaction_game",
                      "periodic_table",
                      "lab_simulation",
                      "puzzle_knowledge"
                    ],
                    "nullable": true,
                    "description": "Danh mục của trò chơi"
                  },
                  "max_score": {
                    "type": "integer",
                    "nullable": true,
                    "description": "Điểm tối đa có thể đạt được"
                  },
                  "time_limit": {
                    "type": "integer",
                    "nullable": true,
                    "description": "Thời gian giới hạn (giây)"
                  },
                  "is_active": {
                    "type": "boolean",
                    "nullable": true,
                    "description": "Trạng thái hoạt động của trò chơi"
                  },
                  "game_settings": {
                    "type": "string",
                    "nullable": true,
                    "example": "{\"question_count\": 10, \"hints_allowed\": true}",
                    "description": "Cài đặt trò chơi dưới dạng chuỗi JSON"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật trò chơi thành công",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                },
                "description": "Cho phép mọi origin (điều chỉnh cho production)"
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string",
                  "example": "PUT, OPTIONS"
                },
                "description": "Các phương thức HTTP được phép"
              },
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string",
                  "example": "Content-Type, Authorization"
                },
                "description": "Các header được phép"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "game_id": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi dữ liệu đầu vào không hợp lệ",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trò chơi",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "headers": {
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Games"],
        "summary": "Xóa trò chơi theo ID",
        "description": "API để xóa một trò chơi dựa trên gameId, yêu cầu xác thực JWT.",
        "operationId": "deleteGame",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID của trò chơi cần xóa",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa trò chơi thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["success"]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Lỗi dữ liệu đầu vào không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập (JWT không hợp lệ)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy trò chơi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Lỗi server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["error"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}